web_app/requirements.txt
boto3==1.19.3
botocore==1.22.3
click==8.0.3
Flask==2.0.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==0.10.0
MarkupSafe==2.0.1
pkg-resources==0.0.0
python-dateutil==2.8.2
python-dotenv==0.19.1
s3transfer==0.5.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
